react_flask/server/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy

# boto3
# flask
# gunicorn
# SQLAlchemy==1.3.16
# Flask-SQLAlchemy==2.4.1
# psycopg2
# jsonifyabsl-py==1.3.0
# appdirs==1.4.4
# apturl==0.5.2
# asttokens==2.0.8
# astunparse==1.6.3
# astutils==0.0.4
# attrs==21.2.0
# autopep8==1.7.1
# Babel==2.8.0
# backcall==0.2.0
# base32hex==1.0.2
# bcrypt==3.2.0
# beautifulsoup4==4.10.0
# beniget==0.4.1
# blinker==1.4
# Bottleneck==1.3.2
# Brlapi==0.8.3
# Brotli==1.0.9
# bs4==0.0.1
# cachetools==5.2.0
# certifi==2020.6.20
# cffi==1.15.1
# chardet==4.0.0
# charset-normalizer==2.1.1
# chronometry==2020.11.12
# cliapp==1.20180812.1
# click==8.0.3
# cloudpickle==2.2.0
# cmdtest==0.32+git
# colorama==0.4.4
# colouration==2020.7.26
# command-not-found==0.3
# cryptography==38.0.1
# cupshelpers==1.0
# cycler==0.11.0
# dask==2022.10.0
# dbus-python==1.2.18
# dd==0.5.7
# debugpy==1.6.3
# decorator==5.1.1
# defer==1.0.6
# defusedxml==0.7.1
# dill==0.3.5.1
# disk==2021.6.29
# distlib==0.3.6
# distro==1.7.0
# distro-info===1.1build1
# docker==6.0.0
# docker-compose==1.29.2
# dockerpty==0.4.1
# docopt==0.6.2
# duplicity==0.8.21
# entrypoints==0.4
# et-xmlfile==1.0.1
# executing==1.1.0
# fasteners==0.14.1
# filelock==3.8.0
# Flask==2.2.2
# Flask-Cors==3.0.10
# flatbuffers==22.9.24
# fonttools==4.29.1
# fs==2.4.12
# fsspec==2022.10.0
# func-timeout==4.3.5
# future==0.18.2
# fuzzywuzzy==0.18.0
# gast==0.4.0
# geographiclib==1.52
# geopy==2.2.0
# google==3.0.0
# google-auth==2.13.0
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.50.0
# gunicorn==20.1.0
# h5py==3.7.0
# html5lib==1.1
# httplib2==0.20.2
# idna==3.3
# imageio==2.22.2
# importlib-metadata==4.6.4
# iniconfig==1.1.1
# ipykernel==6.16.0
# ipython==8.5.0
# isodate==0.6.1
# itsdangerous==2.1.2
# jdcal==1.0
# jedi==0.18.1
# jeepney==0.7.1
# Jinja2==3.0.3
# joblib==1.2.0
# jsonschema==3.2.0
# jupyter-core==4.11.1
# jupyter_client==7.3.5
# keras==2.10.0
# Keras-Preprocessing==1.1.2
# keyring==23.5.0
# kiwisolver==1.3.2
# language-selector==0.1
# launchpadlib==1.10.16
# lazr.restfulclient==0.14.4
# lazr.uri==1.0.6
# libclang==14.0.6
# llvmlite==0.38.0
# locket==1.0.0
# lockfile==0.12.2
# louis==3.20.0
# lxml==4.8.0
# lz4==3.1.3+dfsg
# macaroonbakery==1.3.1
# Mako==1.1.3
# Markdown==3.4.1
# MarkupSafe==2.1.1
# matplotlib==3.5.1
# matplotlib-inline==0.1.6
# monotonic==1.6
# more-itertools==8.10.0
# MouseInfo==0.1.3
# mpmath==0.0.0
# nest-asyncio==1.5.6
# netifaces==0.11.0
# networkx==2.8.7
# nltk==3.7
# numba==0.55.1
# numexpr==2.8.1
# numpy==1.23.3
# oauthlib==3.2.0
# odfpy==1.4.2
# olefile==0.46
# openpyxl==3.0.9
# opt-einsum==3.3.0
# packaging==21.3
# pandas==1.3.5
# paramiko==2.9.3
# parso==0.8.3
# partd==1.3.0
# pdf==2020.11.12
# pdfminer==20191125
# pdfminer.six==20220524
# pdftotree==0.5.0
# pexpect==4.8.0
# pickleshare==0.7.5
# Pillow==9.0.1
# pipenv==2022.11.5
# platformdirs==2.5.2
# pluggy==0.13.0
# ply==3.10
# prompt-toolkit==3.0.31
# protobuf==3.12.4
# psutil==5.9.0
# ptyprocess==0.7.0
# pure-eval==0.2.2
# py==1.10.0
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# PyAutoGUI==0.9.53
# pycairo==1.20.1
# pycodestyle==2.9.1
# pycparser==2.21
# pycryptodome==3.15.0
# pycups==2.0.1
# pydot==1.4.2
# PyGetWindow==0.0.9
# Pygments==2.13.0
# PyGObject==3.42.1
# PyJWT==2.3.0
# pymacaroons==0.13.0
# PyMsgBox==1.0.9
# PyNaCl==1.5.0
# pyparsing==2.4.7
# pyPdf==1.13
# PyPDF2==2.11.1
# pyperclip==1.8.2
# PyRect==0.2.0
# pyRFC3339==1.1
# pyrsistent==0.18.1
# PyScreeze==0.1.28
# pytest==6.2.5
# python-apt==2.3.0+ubuntu2.1
# python-dateutil==2.8.2
# python-debian===0.1.43ubuntu1
# python-dotenv==0.19.2
# python3-xlib==0.15
# pythran==0.10.0
# pytweening==1.0.4
# pytz==2022.1
# PyWavelets==1.4.1
# pywhatkit==5.4
# pyxdg==0.27
# PyYAML==5.4.1
# pyzmq==24.0.1
# ravenclaw==2020.11.12
# rdflib==6.2.0
# regex==2022.9.13
# reportlab==3.6.8
# requests==2.28.1
# requests-oauthlib==1.3.1
# rsa==4.9
# scikit-image==0.19.3
# scikit-learn==1.1.2
# scipy==1.8.0
# SecretStorage==3.3.1
# selectivesearch==0.4
# Send2Trash==1.8.0
# silverware==2020.11.12
# six==1.16.0
# sklearn==0.0
# slytherin==2020.11.12
# soupsieve==2.3.1
# SPARQLWrapper==2.0.0
# ssh-import-id==5.11
# stack-data==0.5.1
# sympy==1.9
# systemd-python==234
# tables==3.7.0
# tabula-py==2.5.1
# tensorboard==2.10.1
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.10.0
# tensorflow-estimator==2.10.0
# tensorflow-io-gcs-filesystem==0.27.0
# termcolor==2.0.1
# texttable==1.6.4
# threadpoolctl==3.1.0
# tifffile==2022.10.10
# toml==0.10.2
# tomli==2.0.1
# toolz==0.12.0
# tornado==6.2
# tqdm==4.64.1
# traitlets==5.4.0
# ttystatus==0.38
# txt==2020.11.12
# typing_extensions==4.4.0
# ubuntu-advantage-tools==27.11.2
# ubuntu-drivers-common==0.0.0
# ufoLib2==0.13.1
# ufw==0.36.1
# unattended-upgrades==0.1
# unicodedata2==14.0.0
# urllib3==1.26.5
# usb-creator==0.3.7
# virtualenv==20.16.6
# virtualenv-clone==0.5.7
# wadllib==1.3.6
# Wand==0.6.10
# wcwidth==0.2.5
# webencodings==0.5.1
# websocket-client==1.4.1
# Werkzeug==2.2.2
# wikipedia==1.4.0
# wrapt==1.14.1
# xdg==5
# xkit==0.0.0
# xlwt==1.3.0
# zipp==1.0.0
